Updated upload to use class and accept flask app
